Lots of changes
Added graphics and card data.
Cards are now instances and not just drawn via GUI.
Added a preliminary way for cards in the hand to "bunch up" and still be accessible.
Cards can be hidden and shown.
Redesigned some card elements to be large for visibility and to allow larger numbers.