Many changes to hands, graveyard, and added same for enemy
ShowHand button's sprite now reacts correctly to whether or not the hand is being shown.
Fixed: Sometimes the graveyard is visible behind the hand while the hand is shown.
Merged the HAND_CONTROLLER and GRAVEYARD_CONTROLLER handling script to work with both (and then again to work for the enemy, so one script covers all 4 scenarios the same way.
Added: a check to see if a graveyard card is a duplicate, and remove the first entry and add it to the end. More of a fix for a potential bug that might never happen under normal gameplay, but worth fixing just in case.
Cards in graveyard can no longer be used as though they were in the hand/on the board (which lead to the previous bug).
Added a click detection script to CONTROLS to determine what object the player is actually trying to click on (the topmost).
Monsters/Creatures can now be played and go to the first free available slot on the board.
Glow effects for highlighting interactive objects like cards works, but needs improving.