Refactor and implementations
Changing the phase of the game now only takes a single click instead of double.
Added a new phase between main1 and combat which is for planning the attack. You can no longer set monsters to attack during the main phase (in theory, in practice you can but it's a bug).

Animated the drawing of cards from deck to hand, still needs work but is a good start.

Refactored the player hand and graveyard. HAND_CONTROLLERs are now just HANDs and are what contain the cards, in a list called "contents". So a PLAYER has a field named "hand" which points to the PLAYER_HAND object, which contains the actual cards in "contents". This is far more logical and simplifies the code a bit.