Cards now slide into place instead of instantly teleporting.
Made it so cards now smoothly slide into place instead of instantly teleporting, but something along the way broke the positioning of card elements (stats, art, etc) while on the board, so fixing that bug will be the next priority, along with noticing that some cards never seem to come to a rest, so that will be investigated.

This could be expanded to placing cards on the board from the hand.