Improvements to stability, added damage counters
Fixed: Some cards never come to a rest
Fixed: First card y position is wrong when it's the only card in the hand.
Fixed (somehow): Skills text appears/disappears when armed for combat
Fixed: Graveyard is behaving oddly

Changes:
Added Damage Counters.
Drawn cards are now drawn on top of everything else.
The hand is now forced to be shown when drawing a card.