Many features, mostly FX-based.
Fixed issues with glowing.
Cards being cast now move to the left of the screen and back to the hand when cancelled.
Added a magic wand effect when casting spells. This wand has particle effects and responds to the mouse movement in a fun way.
Spells now will run an animation when cast. So far only bolt.
Monsters now explode on death.