AI can now cast multiple spells in a row, Bug fixes, some QoL improvements.
Fixed: Monsters in graveyard are not the same size as spells in the graveyard.
Fixed: Enemy cards don't appear faceup when played if the hand isn't shown.
Fixed: You can cast spells on monsters that are in your hand (and probably in the graveyard).
Fixed: Infinite loop when player has no monsters and AI wants to cast a spell.  We didn't have a default action to cast on the player when there's no monsters.
Fixed: We still go to "choose your attacker" when all the monsters have been killed, sounds like they're not being removed from a list upon death.

Done: We need to de-loopify AI_playDestructionSpells() to be more like the summon creature script so we can just iterate manually and that way the waitTime will be effective.
Done: While casting a spell the graveyard and hand should be hidden (unless it targets the grave, etc).

Changed: Removed log_error() and replaced it with log() and debug()