Bug Fixes
Fixed: Heal and bolt don't work on players now.

Fixed: When choosing an attack target the game crashes when clickinng the showhand button.

Fixed: When enemy is drawing a card, the card briefly becomes visible when it reaches the end.

Fixed: You were able to attack an enemy's monster card while it was still in their hand.