Cards are now entirely image based, making them easy to scale.
Changed the system so that when creating a card, we run a script that draws all its contents / elements to a surface and then creates a sprite from that. Each card controls its own font sizes and such, so this method works great for easy scaling and means we don't need to worry about problems with the text moving around or not fitting the box.

I'm a bit concerned about the memory footprint as with only a handful of cards we're already up to about 90 megabytes of memory., so we may need to make the 1:1 scale smaller (maybe 50% of what it is now). Checks will be done when we have more cards.