Finished a basic implementation of the popup message feature.
Feature: Popups are shown when casting a spell or summoning a monster, both for the player and the AI. The AI is paused while the popup is shown and the player can't click on anything else. The popup can be dismissed by clicking anywhere, then the AI will continue with their next action.
To Do: Popups need to be shown for creature attacks, then later for enchantments and ability activation. Title text needs to be centered and the background might gain a border or some graphics to differentiate spells from attacks, etc.

Fixed: The creature card's stats were only updated at end of turn, so now getCardSurface() is ran during the damage() method.