Further work on AI
AI will now attack you with their monsters provided you have no monsters of your own on the field. If you have monsters the AI will not advance yet, so that's next on the TO DO list.

Fixed a logical error where the ai_combat_phase() script was getting called from two places at the same time.

Changed the nextPhase() method so that both player and enemy use the same phases, it's just global.TURN that determines who's in control.