Redesign of Combat AI
Now the AI sort of works. It makes bad decisions and just relentlessly attacks because I haven't implemented any heuristics yet, but it works. Were this a MTG style combat with summoning sickness and choosing defenders it would probably feel much more solid, but there's something to be said for the way things are right now, it's not entirely unfun.