Far more work done on the new mana/resource system.
Feature: Creature attacks have a popup. This was probably already done in a previous update but I forgot to mention it.

Feature: AI now has the same resource costs as the player. Added AI activation of resources for casting spells, summoning, etc.

Feature: Added autotapping of resources which was meant for the AI to tap the right number of cards but also allows the player to autotap too. This was an issue in MTG Shandalar because it wasn't smart enough to make good tapping decisions, so we'll need to be careful about how it works in the future.

Feature: At some point it became possible for the Enemy AI to attack your creatures, I can't recall coding this but I might have done it quickly in a previous update and forgot to mention it, so either way it's a thing now.

Priority To Do: Fix the resource/mana cards lane so that they can be easily clicked and activated. Right now there's a bug that prevents activating some because others are "clicked" instead.