Changes and Improvements
Payment Phase and Combat Phase are automatically skipped when nothing is doable.

Added a constant/macro for the number 15 called CONST_DESTROY_EVENT. This will be used to refer to user_event(15) which handles the destruction of cards from now on. It's a constant because writing event_user(15) doesn't mean anything without a comment explaining it and it's going to be used a lot. It could be an enum either.

Enchantment cards now go to the graveyard when the creature they're on is destroyed.

Changed the variable name "targetType" to "targetAlliance" in Enchantments because it's way more fitting. "targetType" makes you think of Creature, Land, Enchantment etc, which iisn't what the variable was representing.

Bug: The changes not mean things are really acting strangely in the hand and graveyards. Should hopefully be a simple fix but it's definitely a priority.