Fixes to Resources, Changes to Enchantment Cards, a Log system, various smaller tasks and refactoring, some major.
Resources are displayed correctly and click-through no longer occurs.

Enchantment cards are now correctly placing themselves on the target card and are sorted accordingly.

Enchantment cards will deactivate and go to the graveyard when their target creature is destroyed.

Enchantments don't interfere with selecting the monster itself. (had an off-by-one error that was difficult to spot)

Removed Player and Enemy specific Resource Lanes and renamed RESOURCESLOT to RESOURCE_LANE for consistency.

log() will now attempt to open game.log and write to it.