Started implementing card zoom, some refactoring.
Added feature to view cards in their full size, in order to help with reading their data. Might be a littler too big though and obscures vision too much. It also needs to be based on the depth of the cards, so that's next.

Created some helper methods that will help make coding faster in the future:
ds_set_add() and ds_list_remove().

ds_set_add() adds an item to an existing list as though it were a set, checking to see if the value exists already.

ds_list_remove()
Deletes an item from a list by value if it is found.