Fix to hand display depth and
Fix: Cards in the hand are now drawn above resource cards on the playing field as intended.

Fix: When a creature with multiple enchantments was destroyed in combat, the enchantments didn't all go to the graveyard, now they do. It was a case of needing to traverse the loop backwards when deleting items from it.