Added a Marquee, new sprite and behavior for highlighting phase, and functions for controlling these
The marquee is great but it requires a monospaced font, so we'll have to find a decent looking one or make our own (which is the long term plan anyway). The problem with the marquee at the moment is that it's not smooth (especially noticeable at higher scroll speeds), so there'll need to be some kind of abstraction layer to make that possible.

Phase Icons are now highlighted nicely.

Changed how card embiggening works, now it's more like a tooltip. The previous implementation was quite glitchy looking.