Added a button to moving the card to the front slot. Fixed a bug in the AI.
Fixed a bug in the AI that lead to them attacking their own creatures with Bolt.

There is now a symbol that appears when a card is able to move forward to the front lane.

However, sometimes when summoning a creature you don't get a chance to select the slot you want it to be on, so that needs to be fixed.