Fixes to the combat
We can now attack the enemy creatures and opponent directly again.

Changed the Drain ability of the Vampire Bat to only take effect when dealing damage to the opponent directly.

Started implementing an EVENT_LISTENER which should allow us to easily trigger cards when an action is taken, e.g. drawing a card, casting a spell, etc.