Fixes to depths and added new sprites.
Fixed:
Double-clicking on a creature causes is to "activate" and go dark only if it actually has an ability to activate. At the end of turn, "activated" is set to false again.

Change:
Marquee now shows below cards and enchantments thereon.

The Phase Spotlight shows above the Marquee and below phase icons and card slots.
Added alternative sprites for the phase icons so they're grayed out when not their phase.