Implemented a custom testing framework
Game Maker Studio 1.4 doesn't come with any testing framework by default, but it's not a good idea to rely on a human (especially a solo developer) to test things fully, so I decided to implement my own. The current implementation is pretty barebones and will be expanded upon, but the basics are there as far as using assert methods, mocking objects and datastructures, meaningful result info (test name, expected result, actual result) etc. This could theoretically be expanded somewhat into Selenium style automated tests using event_perform() on key objects, which could allow for UI testing, like trying to cast a spell on the enemy directly, does it work or is it broken? (again...)