AI Fixes
Fixed: AI attacks every other turn instead of every turn.

Probably fixed: Sometimes the AI casting Lion's Strength would crash the game. This was almost certainly due to a size-1 error that was returning "undefined" instead of an OutOfBounds error. This reason it's "probably" is because the bug is hard to test, but it seems to be resolved.