Fixed a bug.
Fixed: Refactoring introduced an logic error where clicking on decks wasn't returning their owner as the target like they should. I added a check in the DECK step event to set global.target = owner when clicked outside of the draw phase. Unfortunately to arrive at this fix required a lot of trial and error and so there's some extraneous code in other places that attempted to fix the issue and should be removed.