Fixes and a change to how enemy cards are shown
Fixed:
AI Enchantment cards were organized in the wrong vertical direction.

There was a logic error in organizeEnchantments() which resulted in it always attempting to organize enchantments even where there were none. This lead to me discovering the same logic error in organizeHand() and organizeResources()

Feature: Enemy cards that have enchantments on them now position themselves better, with the top enchantment's y being the y of the card slot.